modules/accessi/backend/requirements.txt
alembic==1.16.5
cryptography==45.0.7
email-validator==2.3.0
fastapi==0.118.0
httpx==0.28.1
openpyxl==3.1.5
pandas==2.3.2
paramiko==3.5.1
PyJWT==2.10.1
psycopg[binary]==3.2.10
pydantic-settings==2.11.0
pytest==8.4.2
python-dotenv==1.1.1
python-multipart==0.0.20
sqlalchemy==2.0.43
uvicorn[standard]==0.37.0
